诗词检索功能和修改格律检测样式

+++ domain/src/main/java/net/yunzhanyi/domain/mapper/TagMapper.java
@@ -45,7 +45,7 @@ public interface TagMapper {
      * @param id id
      * @return {@link Tag}
      */
-    Tag selectByPrimaryKey(Long id);
+    Tag selectByPrimaryKey(Integer id);
 
     /**
      * 更新主键选择性
